feat : 패키지 분리 및 AuthUser 클래스 추가하였습니다.
